add validating types for basic DBus names
This set of types makes it possible to validate strings according to the
requirements for certain "names" as specified in the DBus spec. The sections
"Valid Names", "Valid Object Paths" and "Match Rules" in the spec define the
rules for validaton. The types are:

  * UniqueName - a unique connection name
  * WellKnownName - a unique name alias, aka well-known bus name
  * BusName - either one of the previous two, a "bus address"
  * ObjectPath - a object path as used in the message header
  * InterfaceName - a interface name as used in the message header
  * MemberName - a member name as used in the message header
  * ErrorName - a error name as used in the message header
  * NamespaceName - a special kind of bus name used in match rules

The types are opaque, since they take a C or std:: string and have
an operator std::string().

Signed-off-by: Heiko Hund <heiko@openvpn.net>