extend and overhaul MatchRule type
Change the way MatchRules are constructed. Instead of taking a string
and then parsing it, take the match keys in a structured way, validate
them and generate a match rule string from them.

Added support for missing match rules keys, except "eavesdrop", since that
one is deprecated in favor of the "BecomeMonitor" method.

Signed-off-by: Heiko Hund <heiko@openvpn.net>