redesign DBus type system
The previous approach using boost::any as the generic DBus type had the
downside that any data(type) could be fed into it, no matter if DBus
actually supports the datatype.

Instead, there's now a DBus::Type::Any which behaves more like a
std::variant with some DBus specific extras. It serves as the container
for the others inheriting from DBus::Type and as a generic type the
DBus container types can store.

This type system is more type safe than before, so there's no need to
compare types by their serialized string form, as each can be cast back
to the native C++ type.

The Signature type now validates the signature string passed to it. So,
it is no longer possible to use DictEntry's outside of an Array, like
the standard mandates.

The ObjectPath type makes use of the DBus::ObjectPath "name" and the
validation done therein.

The Array type checks the all Array::add()ed elements have the correct
signature and/or type.

Made the DBus::Type::toString() output a bit more compact and reflect
the container types a bit better by using [], {} and () in the right
places.

Signed-off-by: Heiko Hund <heiko@openvpn.net>