overhaul DBus::Message
Most of things are now done in the the base type DBus::Message (was:
DBus::Message::Base). As a consequence the types representing the four
DBus messages hardly have any code left and their implementation was
moved in with the rest of the message handling code.

The DBus::Message::Header now makes use of the DBus type system and has
support for all header fields. It also got a toString() for logging.

Same goes for the body, a.k.a parameters, which can now be logged for
incoming and outgoing messages using the DBus::Type system.

The MethodCallIdentifier and MethodCallParameters types are now simply
known as Identifier and Paramaeters, since they are used with other
messages than MethodCall as well. MethodCallParametersIn is now the
non-const part of the Parameters type.

Signed-off-by: Heiko Hund <heiko@openvpn.net>