overhaul DBus::AuthenticationProtocol
The class now implements the complete state machine for the DBus
authentication protocol and handles all commands.
It's works completely async.

With little work it should be doable to subclass different auth protocols
from this. However, I'm not sure if there's need for more than what has
been there before, so I left it as it is for now.

Signed-off-by: Heiko Hund <heiko@openvpn.net>