overhaul DBus::MessageProtocol
Sending and receiving of messages is now done asyncronously.

In the receive loop, we first peek at the message to determine its size
and the receive it in one go including file descriptors, potentially.
Received messages are dispatched to their registered handlers, a.k.a
completion tokens, to be dealt with.

Added the DBus::StoredToken helper class to keep a hold of the callers
completion token while we're waiting for a message event it receives
to happen, as ASIO itself currently does not have support for that.

Signed-off-by: Heiko Hund <heiko@openvpn.net>